Release mode builds should have SystemComponent=true
This causes the extension to be hidden from the extension manager dialog,
which is desirable since this is an infrastructure component that users
should mostly never have to deal directly with.

Build script needs to rebuild the VSIX with the flag, and gitlink is run
afterwards as a consequence.

NOTE: the issue with marking VSIXes as such is that they can no longer be
uninstalled via API calls to the ExtensionManager service, meaning the F5
experience becomes pretty much impossible.